feat(#208): fix all qulice suggestions
