+++ src/main/java/com/github/lombrozo/testnames/JUnitExtension.java
@@ -1,0 +1,61 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2022-2023 Volodya
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+package com.github.lombrozo.testnames;
+
+/**
+ * JUnit extension.
+ *
+ * @since 0.1.17
+ */
+public final class JUnitExtension {
+
+    /**
+     * Fully qualified name of the JUnit extension.
+     */
+    private final String name;
+
+    /**
+     * Constructor.
+     * @param klass The JUnit extension class.
+     */
+    public JUnitExtension(final Class<?> klass) {
+        this(klass.getPackage().getName());
+    }
+
+    /**
+     * Constructor.
+     * @param name The fully qualified name of the JUnit extension.
+     */
+    private JUnitExtension(final String name) {
+        this.name = name;
+    }
+
+    /**
+     * Returns true if the JUnit extension is a JUnit extension.
+     * @return True if the JUnit extension is a JUnit extension, false otherwise.
+     */
+    public boolean isJUnitExtension() {
+        return this.name.startsWith("org.junit.jupiter.api.extension");
+    }
+}
