+++ src/main/java/com/github/lombrozo/testnames/javaparser/TestClassJavaParser.java
@@ -26,7 +26,7 @@ package com.github.lombrozo.testnames.javaparser;
 
 import com.github.javaparser.ParseProblemException;
 import com.github.javaparser.StaticJavaParser;
-import com.github.lombrozo.JUnitExtension;
+import com.github.lombrozo.testnames.JUnitExtension;
 import com.github.lombrozo.testnames.TestCase;
 import com.github.lombrozo.testnames.TestClass;
 import java.io.InputStream;
