+++ src/test/java/com/github/lombrozo/testnames/javaparser/ProjectJavaParserTest.java
@@ -37,8 +37,6 @@ import java.util.stream.Collectors;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.AfterAllCallback;
-import org.junit.jupiter.api.extension.ExecutionCondition;
 import org.junit.jupiter.api.io.TempDir;
 
 /**
